+++ tests/test_files.py
@@ -6,15 +6,18 @@ def test_conf_env_var(mocker, tmpdir):
     mocked_load_config = mocker.patch('goodconf._load_config')
     path = tmpdir.join('myapp.json')
     path.write('')
+    class G(GoodConf):
+        class Config:
+            file_env_var = "CONF"
     with env_var('CONF', str(path)):
-        GoodConf(file_env_var='CONF').load()
+        G().load()
     mocked_load_config.assert_called_once_with(str(path))
 
 
 def test_all_env_vars(mocker):
-    mocked_set_values = mocker.patch('goodconf.GoodConf.set_values')
+    mocked_set_values = mocker.patch('goodconf.pydantic.BaseSettings.__init__')
     GoodConf().load()
-    mocked_set_values.assert_called_once_with({})
+    mocked_set_values.assert_called_once_with()
 
 
 def test_provided_file(mocker, tmpdir):
@@ -30,5 +33,8 @@ def test_default_files(mocker, tmpdir):
     path = tmpdir.join('myapp.json')
     path.write('')
     bad_path = tmpdir.join('does-not-exist.json')
-    GoodConf(default_files=[str(bad_path), str(path)]).load()
+    class G(GoodConf):
+        class Config:
+            default_files = [str(bad_path), str(path)]
+    G().load()
     mocked_load_config.assert_called_once_with(str(path))