+++ CHANGES.rst
@@ -2,10 +2,15 @@
 Change Log
 ==========
 
-1.0.1 (unreleased)
+2.0 (unreleased)
 ==================
 
-- Nothing changed yet.
+- Backwards Incompatible: Migrated backend to ``pydantic``.
+  - ``Value`` now mirrors arguments for the `Field function <https://pydantic-docs.helpmanual.io/usage/schema/#field-customisation>`__.
+    ``help`` is now ``description``
+  - ``GoodConf`` is now backed by `BaseSettings <https://pydantic-docs.helpmanual.io/usage/settings/>`__
+    Instead of passing keyword args when instantiating the class, they are now defined on a ``Config`` class on the object
+
 
 
 1.0.0 (18 July 2018)